Renamed StudentMapperNew to StudentMapper
